machineLearning/requirements.txt
Flask==2.2.2
tensorflow==2.12.0
numpy==1.23.5
Pillow==9.3.0
gunicorn==20.1.0